fixes insertion issue regarding mac_addrs and ip_addrs, updates respo… (#71)
* Re-enabled and added compiler flags. Fixed resulting warnings.

* Re-enabled "-Wconversion" (potential type conversion issues).
* Re-enabled "-Wswitch-enum" (ensure all cases explicitly handled).
* Added "-Wnon-virtual-dtor" (base classes with non-virtual destructors).
* Fixed all compiler warnings from enabling/adding those flags.
* Fixed -Wfloat-equal warnings with boost::math::epsilon_difference.

Other code simplifications and cleanup:

* Updated an error-prone std::equal call to safer C++14 version.
* Converted some manual floating-point formatting over to boost::format.

* Converted old SQL "WHERE NOT EXISTS" workaround to new "ON CONFLICT" syntax.

* Added multiple UNIQUE partial indexes to tables without PRIMARY KEYs.

* Fixed compiler warnings in the unit tests.

* Added type-checking DOMAINs `PortNumber` and `VlanNumber`.

Switched `port` and `vlan` columns to use the new DOMAIN types.
This also allowed removing per-instance CHECK clauses.

* Added documentation comments. Reduced VlanNumber from INT to SMALLINT.

* Added floating point inequality test cases to extraWeight unit tests.

* Added more floating point inequality test cases to extraWeight unit tests.

* fixes insertion issue regarding mac_addrs and ip_addrs, updates responding now

* change boolean OR comparison to GREATEST, prevents NULL creep due to behavior of (NULL OR FALSE)

* initial test of expressional indexes

* convert targeted unique indexes to expressional

* clean out trailing whitespace

* using hash chain as expresional index now

+++ datastore/exporters/nmdb-export-query/WriterContext.cpp
@@ -86,7 +86,7 @@ WriterContext::addQueryInfo(std::string _queryInfo, std::string _query)
 }
 
 void
-WriterContext::addColumn(std::string _colName, float _width)
+WriterContext::addColumn(std::string _colName, double _width)
 {
   columns.push_back(_colName);
   columnWidths.emplace(_colName, _width);
